% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{assignTax}
\alias{assignTax}
\title{Assign taxonomy functions}
\usage{
assignTax(
  analysis_setup,
  asv_abund_matrix,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = FALSE,
  retrieve_files = FALSE,
  db_rps10 = "oomycetedb.fasta",
  db_its = "fungidb.fasta",
  db_16s = "bacteriadb.fasta",
  db_other = "otherdb.fasta",
  overwrite_existing = FALSE
)
}
\arguments{
\item{analysis_setup}{A list containing directory paths and data tables, produced by the \code{prepare_reads} function.}

\item{asv_abund_matrix}{ASV abundance matrix.}

\item{tryRC}{Whether to try reverse complementing sequences during taxonomy assignment.}

\item{verbose}{Logical, indicating whether to display verbose output.}

\item{multithread}{Logical, indicating whether to use multithreading.}

\item{retrieve_files}{Specify TRUE/FALSE whether to copy files from the temp directory to the output directory.}

\item{overwrite_existing}{Logical, indicating whether to remove or overwrite existing files and directories from previous runs.}
}
\value{
Taxonomic assignments of each unique ASV sequence
}
\description{
Assign taxonomy functions
}
\examples{
Assign taxonomies to ASVs on a per barcode basis
prepare_reads(maxN = 0, data_directory = "~/demulticoder/inst/extdata", output_directory = "~/testing_package", tempdir_id = "run1", overwrite_existing = TRUE)
cut_trim(analysis_setup,cutadapt_path="/opt/homebrew/bin/cutadapt", overwrite_existing = TRUE)
make_asv_abund_matrix(analysis_setup, overwrite_existing = TRUE)
assignTax(analysis_setup,asv_abund_matrix, retrieve_files=TRUE, overwrite_existing=TRUE)
}
