% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{assignTax_as_char}
\alias{assignTax_as_char}
\title{Combine taxonomic assignments and bootstrap values for each locus into single falsification vector}
\usage{
assignTax_as_char(tax_results)
}
\description{
Combine taxonomic assignments and bootstrap values for each locus into single falsification vector
}
\keyword{internal}
