% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertASVMatrix.R
\name{asvmatrix_to_taxmap}
\alias{asvmatrix_to_taxmap}
\title{Filter ASV abundance matrix and convert to taxmap object}
\usage{
asvmatrix_to_taxmap(
  min_read_depth = 0,
  minimum_bootstrap = 50,
  pid_species = 0,
  pid_genus = 0,
  pid_family = 0
)
}
\arguments{
\item{min_read_depth}{Threshold for ASVs to remove if number of
reads is less than this value across all samples-todo check on this}

\item{minimum_bootstrap}{Threshold for bootstrap support value
for taxonomic assignments. Below designated minimum bootstrap
threshold, taxnomoic assignments will be set to N/A.}

\item{pid_species}{If percent identity is below this value, at the species
level, taxonomic assignment will be set to N/A}

\item{pid_genus}{If percent identity is below this value, at the genus
level, taxonomic assignment will be set to N/A}

\item{pid_family}{If percent identity is below this value, at the family
level, taxonomic assignment will be set to N/A}

\item{asv_abund_matrix}{The final ASV abundance matrix}
}
\value{
ASV matrix converted to taxmap object
}
\description{
Filter ASV abundance matrix and convert to taxmap object
}
\examples{
directory_path<-"~/rps10package/raw_data/rps10_ITS"
primer_path <-file.path(directory_path, "primer_info.csv")
metadata_path <-file.path(directory_path,"metadata.csv")
cutadapt_path<-"/opt/homebrew/bin/cutadapt"
data_tables <-
prepare_reads(
directory_path,
primer_path,
metadata_path,
maxN = 0,
)
cut_trim(
directory_path,
cutadapt_path,
verbose = TRUE,
maxEE = 2,
truncQ = 5,
minLen = 200,
maxLen = 297,
minCutadaptlength = 50
)
asv_abund_matrix <-
make_asv_abund_matrix(
directory_path,
minOverlap = 15,
maxMismatch = 2,
verbose = TRUE
)
summary <- assignTax(
directory_path,
data_tables,
asv_abund_matrix,
multithread = TRUE,
barcode = "rps10"
)
obj_dada <-
asvmatrix_to_taxmap(
asv_abund_matrix,
min_read_depth = 10,
minimum_bootstrap = 75
)
}
