% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_databases.R
\name{create_ref_database}
\alias{create_ref_database}
\title{Prepare reference databases folder if it doesn't exist.}
\usage{
create_ref_database(directory_path)
}
\arguments{
\item{directory_path}{A path to the intermediate folder}
}
\value{
A new directory called reference_databases
}
\description{
Prepare reference databases folder if it doesn't exist.
}
