% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutPrimers_TrimReads.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Wrapper function for filterAndTrim function from DADA2 after primer removal}
\usage{
filter_and_trim(
  directory_path,
  directory_path_temp,
  cutadapt_data_barcode,
  barcode_params,
  barcode
)
}
\arguments{
\item{directory_path}{The path to the directory containing the fastq,
metadata, and primersinfo_params files}

\item{cutadapt_data_barcode}{directory_data folder with trimmed and filtered reads for each sample}
}
\value{
Filtered and trimmed reads
}
\description{
Wrapper function for filterAndTrim function from DADA2 after primer removal
}
\keyword{internal}
