% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{format_abund_matrix}
\alias{format_abund_matrix}
\title{Format ASV abundance matrix}
\usage{
format_abund_matrix(
  data_tables,
  asv_abund_matrix,
  seq_tax_asv,
  output_directory_path,
  locus
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, primer sequences}

\item{asv_abund_matrix}{An abundance matrix containing amplified sequence
variants}

\item{seq_tax_asv}{An amplified sequence variants matrix with taxonomic
information}
}
\description{
Format ASV abundance matrix
}
\keyword{internal}
