% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_database}
\alias{format_database}
\title{General functions to format user-specified databases}
\usage{
format_database(
  data_tables,
  data_path,
  output_directory_path,
  temp_directory_path,
  barcode,
  db_its,
  db_rps10,
  db_16S,
  db_other1,
  db_other2
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, primer sequences}

\item{data_path}{Path to the data directory}

\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{temp_directory_path}{User-defined temporary directory to place reads
throughout the workflow metadata, and primer_info files}

\item{barcode}{The barcode for which the database should be formatted}
}
\value{
A formatted database based on the specified barcode type
}
\description{
General functions to format user-specified databases
}
\keyword{internal}
