% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_databases.R
\name{format_database_16s}
\alias{format_database_16s}
\title{An 16S database that has modified headers and is output in the reference_databases folder.}
\usage{
format_database_16s(directory_path, database_16s)
}
\arguments{
\item{directory_path}{A path to a directory that contains raw data}

\item{database_its}{The name of the database}
}
\value{
A rps10 database that has modified headers and is output in the reference_databases folder.
}
\description{
An 16S database that has modified headers and is output in the reference_databases folder.
}
