% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_databases.R
\name{format_database_its}
\alias{format_database_its}
\title{An ITS database that has modified headers and is output in the reference_databases folder.}
\usage{
format_database_its(directory_path, directory_path_temp, database_its)
}
\arguments{
\item{directory_path}{The path to the directory containing the fastq,
metadata, and primer_info files}

\item{directory_path_temp}{User-defined temporary directory to place reads throughout the workflow
metadata, and primer_info files}

\item{database_its}{The name of the database}
}
\value{
An ITS database that has modified headers and is output in the reference_databases folder.
}
\description{
An ITS database that has modified headers and is output in the reference_databases folder.
}
\keyword{internal}
