% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_databases.R
\name{format_database_rps10}
\alias{format_database_rps10}
\title{Create modified reference rps10 database for downstream analysis}
\usage{
format_database_rps10(directory_path, directory_path_temp, database_rps10)
}
\arguments{
\item{directory_path}{The path to the directory containing the fastq,
metadata, and primer_info files}

\item{directory_path_temp}{User-defined temporary directory to place reads throughout the workflow
metadata, and primer_info files}

\item{database_rps10}{The name of the database}
}
\value{
A rps10 database that has modified headers and is output in the reference_databases folder.
}
\description{
Create modified reference rps10 database for downstream analysis
}
\keyword{internal}
