% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_db_its}
\alias{format_db_its}
\title{An ITS database that has modified headers and is output in the
reference_databases folder}
\usage{
format_db_its(
  data_tables,
  data_path,
  output_directory_path,
  temp_directory_path,
  db_its
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, and metabarcode information with associated primer sequences}

\item{data_path}{Path to the data directory}

\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{temp_directory_path}{User-defined temporary directory to output unfiltered, trimmed, and filtered read directories throughout the workflow}

\item{db_its}{The UNITE ITS reference database provided by the user}
}
\value{
The UNITE ITS database with modified headers
}
\description{
An ITS database that has modified headers and is output in the
reference_databases folder
}
\keyword{internal}
