% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_databases.R
\name{format_db_other}
\alias{format_db_other}
\title{An other, user-specified database that is initially in UNITE fungal db format}
\usage{
format_db_other(analysis_setup, db_other)
}
\arguments{
\item{db_other}{The name of the database}

\item{directory_path}{The path to the directory containing the fastq,
metadata, and primer_info files}

\item{directory_path_temp}{User-defined temporary directory to place reads throughout the workflow
metadata, and primer_info files}
}
\value{
An other database that has modified headers and is output in the reference_databases folder.
}
\description{
An other, user-specified database that is initially in UNITE fungal db format
}
\keyword{internal}
