% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_db_rps10}
\alias{format_db_rps10}
\title{Create modified reference rps10 database for downstream analysis}
\usage{
format_db_rps10(
  data_tables,
  data_path,
  output_directory_path,
  temp_directory_path,
  db_rps10
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, and metabarcode information with associated primer sequences}

\item{data_path}{Path to the data directory}

\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{temp_directory_path}{User-defined temporary directory to output unfiltered, trimmed, and filtered read directories throughout the workflow}

\item{db_rps10}{The oomyceteDB rps10 reference database provided by the user}
}
\value{
The oomyceteDB database with modified headers
}
\description{
Create modified reference rps10 database for downstream analysis
}
\keyword{internal}
