% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{get_align_pid}
\alias{get_align_pid}
\title{Align ASV sequences to reference sequences from database to get percent ID.}
\usage{
get_align_pid(ref, asv)
}
\description{
Align ASV sequences to reference sequences from database to get percent ID.
}
