% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_primers_trim_reads.R
\name{get_post_trim_hits}
\alias{get_post_trim_hits}
\title{Get primer counts for reach sample after primer removal and trimming steps}
\usage{
get_post_trim_hits(primer_data, cutadapt_data, output_directory_path)
}
\arguments{
\item{primer_data}{Primer \code{data.frame} created using the orient_primers function to parse information on forward and reverse primer sequences.}

\item{cutadapt_data}{FASTQ read files trimmed of primers}

\item{output_directory_path}{The path to the directory where resulting files
are output}
}
\value{
Table of read counts across each sample
}
\description{
Get primer counts for reach sample after primer removal and trimming steps
}
\keyword{internal}
