% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{get_read_counts}
\alias{get_read_counts}
\title{Final inventory of read counts after each step from input to removal of chimeras. This function deals with if you have more than one sample. TODO optimize for one sample}
\usage{
get_read_counts(asv_abund_matrix)
}
\arguments{
\item{asv_abund_matrix}{An abundance matrix containing amplified sequence variants}
}
\description{
Final inventory of read counts after each step from input to removal of chimeras. This function deals with if you have more than one sample. TODO optimize for one sample
}
\keyword{internal}
