% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{get_ref_seq}
\alias{get_ref_seq}
\title{Align ASV sequences to reference sequences from database to get percent ID.
Start by retrieving reference sequences.}
\usage{
get_ref_seq(tax_results, db)
}
\arguments{
\item{tax_results}{The dataframe containing taxonomic assignments}

\item{db}{The reference database}
}
\description{
Align ASV sequences to reference sequences from database to get percent ID.
Start by retrieving reference sequences.
}
\keyword{internal}
