% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asv_abund_matrix.R
\name{infer_asv_command}
\alias{infer_asv_command}
\title{Function to infer ASVs, for multiple loci}
\usage{
infer_asv_command(
  output_directory_path,
  temp_directory_path,
  data_tables,
  barcode_params,
  barcode
)
}
\arguments{
\item{output_directory_path}{The path to the directory where resulting files are output}

\item{data_tables}{The data tables containing the paths to read files, metadata, primer sequences}

\item{denoised_data_path}{Path to saved intermediate denoised data}
}
\description{
Function to infer ASVs, for multiple loci
}
\keyword{internal}
