% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asv_abund_matrix.R
\name{infer_asvs}
\alias{infer_asvs}
\title{Core DADA2 function to learn errors and infer ASVs}
\usage{
infer_asvs(
  my_direction,
  my_primer_pair_id,
  barcode_params,
  output_directory_path
)
}
\arguments{
\item{my_direction}{Location of read files and metadata file}

\item{my_primer_pair_id}{The the specific barcode id}

\item{output_directory_path}{The path to the directory containing the fastq,
metadata, and primerinfo_params files}
}
\value{
asv_data
}
\description{
Core DADA2 function to learn errors and infer ASVs
}
\keyword{internal}
