% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asv_abund_matrix.R
\name{make_abund_matrix}
\alias{make_abund_matrix}
\title{Quality filtering to remove chimeras and short sequences}
\usage{
make_abund_matrix(
  raw_seqtab,
  temp_directory_path,
  barcode_params = barcode_params,
  barcode
)
}
\arguments{
\item{raw_seqtab}{An RData file containing intermediate read data before chimeras were removed}
}
\value{
asv_abund_matrix The returned final ASV abundance matrix
}
\description{
Quality filtering to remove chimeras and short sequences
}
\keyword{internal}
