% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Files_Primers.R
\name{make_primer_hit_plot}
\alias{make_primer_hit_plot}
\title{Make a barplot of primers identified on reads}
\usage{
make_primer_hit_plot(primer_hits, fastq_data, directory_path, plot_name)
}
\arguments{
\item{primer_hits}{A number of reads in which the primer is found}

\item{fastq_data}{A tibble with the fastq file paths, the direction of the sequences, and names of sequences}

\item{plot_name}{A filename under which a PDF file of the plot will be saved as}
}
\description{
Make a barplot of primers identified on reads
}
