% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{make_seqhist}
\alias{make_seqhist}
\title{Plots a histogram of read length counts of all sequences within the ASV matrix}
\usage{
make_seqhist(asv_abund_matrix)
}
\description{
Plots a histogram of read length counts of all sequences within the ASV matrix
}
