% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{makeSeqtab}
\alias{makeSeqtab}
\title{Make ASV sequence matrix}
\usage{
makeSeqtab(merged_reads, orderBy = "abundance")
}
\arguments{
\item{merged_reads}{}
}
\description{
Make ASV sequence matrix
}
