% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeASVAbundMatrix.R
\name{merge_reads_command}
\alias{merge_reads_command}
\title{Merge forward and reverse reads}
\usage{
merge_reads_command(
  directory_path,
  directory_path_temp,
  barcode_params,
  barcode
)
}
\arguments{
\item{directory_path}{A path to the intermediate folder and directory}

\item{merged_read_data_path}{Path to R data file containing merged read data}
}
\value{
merged_reads Intermediate merged read R data file
}
\description{
Merge forward and reverse reads
}
\keyword{internal}
