% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{orient_primers}
\alias{orient_primers}
\title{Take in user's primers and creates the complement, reverse,
reverse complement of primers in one tibble}
\usage{
orient_primers(primer_path)
}
\arguments{
\item{primer_path}{a path to the csv file that holds the primer
information}
}
\value{
A data frame with oriented primer information
}
\description{
Take in user's primers and creates the complement, reverse,
reverse complement of primers in one tibble
}
\keyword{internal}
