% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutPrimers_TrimReads.R
\name{plot_post_trim_qc}
\alias{plot_post_trim_qc}
\title{Wrapper script for plotQualityProfile after trim steps and primer removal.}
\usage{
plot_post_trim_qc(cutadapt_data, directory_path, n = 5e+05)
}
\arguments{
\item{cutadapt_data}{directory_data folder with trimmed and filtered reads for each sample}

\item{directory_path}{The path to the directory containing the fastq,
metadata, and primer_info files}

\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}
}
\value{
Quality profiles of reads after primer trimming
}
\description{
Wrapper script for plotQualityProfile after trim steps and primer removal.
}
\keyword{internal}
