% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_primers_trim_reads.R
\name{plot_qc}
\alias{plot_qc}
\title{Wrapper function for plotQualityProfile function}
\usage{
plot_qc(cutadapt_data, directory_path, n = 5e+05)
}
\arguments{
\item{cutadapt_data}{directory_data folder with trimmed and filtered reads for each sample}

\item{directory_path}{The path to the directory containing the fastq,
metadata, and primers and params files}

\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}
}
\value{
Dada2 wrapper function for making quality profiles for each sample
}
\description{
Wrapper function for plotQualityProfile function
}
\keyword{internal}
