% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{prep_abund_matrix}
\alias{prep_abund_matrix}
\title{Prepare final ASV abundance matrix}
\usage{
prep_abund_matrix(cutadapt_data, asv_abund_matrix, data_tables, metabarcode)
}
\arguments{
\item{asv_abund_matrix}{The final abundance matrix containing amplified sequence variants}

\item{metabarcode}{The metabarcode used throughout the workflow (applicable options: 'rps10', 'its', 'r16S', 'other1', other2')}
}
\description{
Prepare final ASV abundance matrix
}
\keyword{internal}
