% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{prepare_metadata_table}
\alias{prepare_metadata_table}
\title{Read in the metadata from user and combine it with the primer data.
Included in a larger function prepare_reads.}
\usage{
prepare_metadata_table(metadata_path, primer_data)
}
\arguments{
\item{metadata_path}{A path to the metadata .csv file}

\item{primer_data}{The primer data tibble created in orient_primers
function}

\item{directory_path}{Folder path of read files and metadata file}
}
\value{
A metadata table containing the concatenated metadata and primer
data
}
\description{
Read in the metadata from user and combine it with the primer data.
Included in a larger function prepare_reads.
}
\keyword{internal}
