% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reads_count_primers.R
\name{primer_check}
\alias{primer_check}
\title{Matching Order Primer Check}
\usage{
primer_check(fastq_data)
}
\arguments{
\item{fastq_data}{A data frame with the FASTQ file paths, the direction of
the sequences, and names of sequences}
}
\value{
None
}
\description{
Matching Order Primer Check
}
\keyword{internal}
