% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{process_single_barcode}
\alias{process_single_barcode}
\title{Process the information from an ASV abundance matrix to run DADA2 for single barcode}
\usage{
process_single_barcode(
  data_tables,
  directory_path_temp,
  directory_path,
  asv_abund_matrix,
  tryRC = tryRC,
  outputBootstraps = outputBootstraps,
  minBoot = minBoot,
  verbose = verbose,
  multithread = multithread,
  locus = barcode
)
}
\arguments{
\item{data_tables}{A list containing data modified by cutadapt, primer data, FASTQ data, and concatenated metadata and primer data}

\item{asv_abund_matrix}{An abundance matrix containing amplified sequence variants}

\item{tryRC}{(Optional). Default FALSE. 
If TRUE, the reverse-complement of each sequences will be used for classification if it is a better match to the reference
sequences than the forward sequence.}

\item{outputBootstraps}{(Optional). Default FALSE.
If TRUE, bootstrap values will be retained in an integer matrix. A named list containing the assigned taxonomies (named "taxa") 
and the bootstrap values (named "boot") will be returned. Minimum bootstrap confidence filtering still takes place,
to see full taxonomy set minBoot=0}

\item{minBoot}{(Optional). Default 50. 
The minimum bootstrap confidence for assigning a taxonomic level.}

\item{verbose}{(Optional). Default FALSE.
If TRUE, print status to standard output.}

\item{multithread}{(Optional). Default is FALSE.
If TRUE, multithreading is enabled and the number of available threads is automatically determined.   
If an integer is provided, the number of threads to use is set by passing the argument on to
\code{\link{setThreadOptions}}.}
}
\description{
Process the information from an ASV abundance matrix to run DADA2 for single barcode
}
\keyword{internal}
