% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{read_fastq}
\alias{read_fastq}
\title{Takes in the fastq files from the user and creates a tibble with
the paths to files that will be created and used in the future.
Included in a larger 'read_prefilt_fastq' function}
\usage{
read_fastq(directory_path)
}
\arguments{
\item{directory_path}{Folder path of read files and metadata file}
}
\value{
A tibble with the fastq file paths, the direction of the
}
\description{
Takes in the fastq files from the user and creates a tibble with
the paths to files that will be created and used in the future.
Included in a larger 'read_prefilt_fastq' function
}
\keyword{and}
\keyword{internal}
\keyword{names}
\keyword{of}
\keyword{sequences}
\keyword{sequences,}
