% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reads_count_primers.R
\name{read_parameters}
\alias{read_parameters}
\title{Take in user's DADA2 parameters and make a dataframe for downstream steps}
\usage{
read_parameters(primers_params_path)
}
\arguments{
\item{primers_params_path}{A path to the CSV file that holds the primer
information.}
}
\value{
A data frame with information on the DADA2 parameters.
}
\description{
Take in user's DADA2 parameters and make a dataframe for downstream steps
}
\keyword{internal}
