% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_Files_Primers.R
\name{read_prefilt_fastq}
\alias{read_prefilt_fastq}
\title{A function for calling read_fastq, primer_check, and remove_ns functions. This will process and edit the FASTQ and make them ready for the trimming of primers with Cutadapt. Part of a larger 'prepare_reads' function.}
\usage{
read_prefilt_fastq(directory_path, maxN = 0, multithread = FALSE)
}
\arguments{
\item{directory_path}{A path to a directory that contains raw data}

\item{maxN}{(Optional). Default 0.
After truncation, sequences with more than \code{maxN} Ns will be discarded. 
Note that \code{\link[dada2]{dada}} does not allow Ns.}

\item{multithread}{(Optional). Default is FALSE.
 If TRUE, input files are filtered in parallel via \code{\link[parallel]{mclapply}}.
 If an integer is provided, it is passed to the \code{mc.cores} argument of \code{\link[parallel]{mclapply}}.
 Note that the parallelization here is by forking, and each process is loading another fastq file into
 memory. This option is ignored in Windows, as Windows does not support forking, with \code{mc.cores} set to 1.
If memory is an issue, execute in a clean environment and reduce the chunk size \code{n} and/or
 the number of threads.}

\item{metadata}{A metadata containing the concatenated metadata and primer data}
}
\description{
A function for calling read_fastq, primer_check, and remove_ns functions. This will process and edit the FASTQ and make them ready for the trimming of primers with Cutadapt. Part of a larger 'prepare_reads' function.
}
\examples{
 Part of the function prepare_reads. fastq_data <- read_prefilt_fastq(directory_path, maxN = maxN, multithread = multithread)
}
