% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_primers_trim_reads.R
\name{run_cutadapt}
\alias{run_cutadapt}
\title{Core function for running 'Cutadapt'}
\usage{
run_cutadapt(
  cutadapt_path,
  cutadapt_data_barcode,
  barcode_params,
  minCutadaptlength
)
}
\arguments{
\item{cutadapt_path}{A path to the 'Cutadapt' program.}

\item{minCutadaptlength}{Read lengths that are lower than this threshold will
be discarded. Default is 0.}
}
\value{
Trimmed read.
}
\description{
Core function for running 'Cutadapt'
}
\keyword{internal}
