% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{separate_abund_matrix}
\alias{separate_abund_matrix}
\title{Separate abundance matrices}
\usage{
separate_abund_matrix(
  abund_asv_barcode2,
  abund_asv_barcode1,
  directory_path,
  asv_abund_matrix
)
}
\arguments{
\item{directory_path}{A path to the intermediate folder and directory}
}
\description{
Separate abundance matrices
}
\keyword{internal}
