% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{setup_paths}
\alias{setup_paths}
\title{Set Up Paths for Analysis}
\usage{
setup_paths(
  directory_path = "inst/extdata",
  temp_path_run_no = "run1",
  primer_file = "primer_info.csv",
  metadata_file = "metadata.csv",
  cutadapt_path = "/opt/homebrew/bin/cutadapt"
)
}
\arguments{
\item{directory_path}{A character string specifying the directory path. Default is "inst/extdata".}

\item{temp_path_run_no}{A character string specifying temporary path run number. Default is "run1".}

\item{primer_file}{A character string specifying the filename of the primer information. Default is "primer_info.csv".}

\item{metadata_file}{A character string specifying the filename of the metadata. Default is "metadata.csv".}

\item{cutadapt_path}{A character string specifying the path to the cutadapt tool. Default is "/usr/bin/cutadapt".}
}
\value{
A list containing paths for the directory path ("directory_path"), temporary path run number ("temp_path_run_no"), primer file ("primer_path"), and metadata file ("metadata_path").
}
\description{
This function sets up the necessary paths for the analysis, based on user input or default settings.
The function returns a list containing paths for the primer file, metadata file, and a temporary directory.
}
\examples{
paths <- setup_paths(directory_path = "path/to/my/data")
print(paths$primer_path)


}
