% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{taxmap_to_phyloseq}
\alias{taxmap_to_phyloseq}
\title{Convert taxmap object to Phyloseq object (metacoder wrapper function)}
\usage{
taxmap_to_phyloseq(obj_dada)
}
\value{
A phyloseq object
}
\description{
Convert taxmap object to Phyloseq object (metacoder wrapper function)
}
