% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{add_pid_to_tax}
\alias{add_pid_to_tax}
\title{Add PID and bootstrap values to tax result.}
\usage{
add_pid_to_tax(tax_results, asv_pid)
}
\description{
Add PID and bootstrap values to tax result.
}
\keyword{internal}
