% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{assign_taxonomyDada2}
\alias{assign_taxonomyDada2}
\title{Assign taxonomy}
\usage{
assign_taxonomyDada2(
  asv_abund_matrix,
  temp_directory_path,
  minBoot = 0,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = TRUE,
  metabarcode = "barcode",
  barcode_params
)
}
\arguments{
\item{asv_abund_matrix}{The final abundance matrix containing amplified sequence variants}

\item{temp_directory_path}{The temporary directory path}

\item{minBoot}{Minimum bootstrap value for taxonomy assignment (default 0)}

\item{tryRC}{Try reverse complement (default FALSE)}

\item{verbose}{Print additional information (default FALSE)}

\item{multithread}{Use multiple threads (default TRUE)}

\item{metabarcode}{The metabarcode for taxonomy assignment (e.g., rps10, other1, other2)}
}
\description{
Assign taxonomy
}
\keyword{internal}
