% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{assign_taxonomyDada2}
\alias{assign_taxonomyDada2}
\title{Assign taxonomy}
\usage{
assign_taxonomyDada2(
  asv_abund_matrix,
  temp_directory_path,
  metabarcode = "barcode",
  barcode_params
)
}
\arguments{
\item{asv_abund_matrix}{The final abundance matrix containing amplified sequence variants.}

\item{temp_directory_path}{User-defined temporary directory to output unfiltered, trimmed, and filtered read directories throughout the workflow}

\item{metabarcode}{The metabarcode used throughout the workflow (applicable options: 'rps10', 'its', 'r16S', 'other1', other2')}
}
\description{
Assign taxonomy
}
\keyword{internal}
