% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{assignTax}
\alias{assignTax}
\title{Assign rps10 and ITS taxonomy}
\usage{
assignTax(
  directory_path,
  database_rps10,
  database_its,
  data_tables,
  asv_abund_matrix,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = FALSE,
  is_ITS = FALSE
)
}
\arguments{
\item{tryRC}{(Optional). Default FALSE. 
If TRUE, the reverse-complement of each sequences will be used for classification if it is a better match to the reference
sequences than the forward sequence.}

\item{verbose}{(Optional). Default FALSE.
If TRUE, print status to standard output.}

\item{multithread}{(Optional). Default is FALSE.
If TRUE, multithreading is enabled and the number of available threads is automatically determined.   
If an integer is provided, the number of threads to use is set by passing the argument on to
\code{\link{setThreadOptions}}.}

\item{is_ITS}{TRUE or FALSE}
}
\description{
Assign rps10 and ITS taxonomy
}
