% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main_functions.R
\name{asvmatrix_to_taxmap}
\alias{asvmatrix_to_taxmap}
\title{Filter ASV abundance matrix and convert to taxmap object}
\usage{
asvmatrix_to_taxmap(
  min_read_depth = 0,
  minimum_bootstrap = 50,
  pid_species = 0,
  pid_genus = 0,
  pid_family = 0
)
}
\description{
Filter ASV abundance matrix and convert to taxmap object
}
