% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_asv_matrix.R
\name{convert_asv_matrix_to_objs}
\alias{convert_asv_matrix_to_objs}
\title{Filter ASV abundance matrix and convert to taxmap object}
\usage{
convert_asv_matrix_to_objs(
  min_read_depth = 0,
  minimum_bootstrap = 0,
  save_outputs = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{min_read_depth}{Threshold for ASVs to remove if number of
reads is less than this value across all samples-todo check on this}

\item{minimum_bootstrap}{Threshold for bootstrap support value
for taxonomic assignments. Below designated minimum bootstrap
threshold, taxnomoic assignments will be set to N/A}

\item{save_outputs}{Logical, indicating whether to save the taxmap object. Default is FALSE.}
}
\value{
ASV matrix converted to taxmap object
}
\description{
Filter ASV abundance matrix and convert to taxmap object
}
\examples{
Convert final matrix to taxmap and phyloseq objects for downstream analysis steps
prepare_reads(maxN = 0, data_directory = "~/demulticoder/inst/extdata", output_directory = "~/testing_package", tempdir_id = "run1", overwrite_existing = TRUE)
cut_trim(analysis_setup,cutadapt_path="/opt/homebrew/bin/cutadapt", overwrite_existing = TRUE)
make_asv_abund_matrix(analysis_setup, overwrite_existing = TRUE)
assign_tax(analysis_setup,asv_abund_matrix, retrieve_files=TRUE, overwrite_existing=TRUE
convert_asv_matrix_to_objs(save_outputs=TRUE)
}
