% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeASVAbundMatrix.R
\name{countOverlap}
\alias{countOverlap}
\title{Count overlap to see how well the reads were merged}
\usage{
countOverlap(merged_reads, data_tables, directory_path, barcode)
}
\arguments{
\item{merged_reads}{Intermediate merged read R data file}

\item{directory_path}{Directory path to save the plot}
}
\value{
A plot describing how well reads merged and information on overlap between reads
Count overlap to see how well the reads were merged

A plot describing how well reads merged and information on overlap between reads
}
\description{
Count overlap to see how well the reads were merged
}
