% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{countOverlap}
\alias{countOverlap}
\title{Count overlap to see how well the reads were merged}
\usage{
countOverlap(merged_reads)
}
\description{
Count overlap to see how well the reads were merged
}
