% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeASVAbundMatrix.R
\name{countOverlap}
\alias{countOverlap}
\title{Count overlap to see how well the reads were merged}
\usage{
countOverlap(merged_reads, directory_path)
}
\description{
Count overlap to see how well the reads were merged
}
