% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asv_abund_matrix.R
\name{countOverlap}
\alias{countOverlap}
\title{Count overlap to see how well the reads were merged}
\usage{
countOverlap(data_tables, merged_reads, barcode, output_directory_path)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, primer sequences}

\item{merged_reads}{Intermediate merged read R data file}

\item{barcode}{The barcode used for the analysis}

\item{output_directory_path}{The path to the directory where resulting files
are output}
}
\value{
A plot describing how well reads merged and information on overlap
between reads
}
\description{
Count overlap to see how well the reads were merged
}
\keyword{internal}
