% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_primers_trim_reads.R
\name{filter_and_trim}
\alias{filter_and_trim}
\title{Wrapper function for filterAndTrim function from 'dada2', to be used after
primer trimming}
\usage{
filter_and_trim(
  output_directory_path,
  temp_directory_path,
  cutadapt_data_barcode,
  barcode_params,
  barcode
)
}
\arguments{
\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{cutadapt_data_barcode}{Metabarcode-specific FASTQ read files trimmed of primers}
}
\value{
Filtered and trimmed reads
}
\description{
Wrapper function for filterAndTrim function from 'dada2', to be used after
primer trimming
}
\keyword{internal}
