% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_database}
\alias{format_database}
\title{General functions to format user-specified databases}
\usage{
format_database(
  analysis_setup,
  barcode,
  db_its,
  db_rps10,
  db_16s,
  db_other1,
  db_other2
)
}
\arguments{
\item{analysis_setup}{An object containing directory paths and data tables,
produced by the \code{prepare_reads} function}

\item{barcode}{The barcode for which the database should be formatted}
}
\value{
A formatted database based on the specified barcode type
}
\description{
General functions to format user-specified databases
}
\keyword{internal}
