% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Format_databases.R
\name{format_database_its}
\alias{format_database_its}
\title{An ITS database that has modified headers and is output in the reference_databases folder.}
\usage{
format_database_its(directory_path, database_its)

format_database_its(directory_path, database_its)
}
\arguments{
\item{directory_path}{A path to a directory that contains raw data}

\item{database_its}{The name of the database}
}
\value{
An ITS database that has modified headers and is output in the reference_databases folder.

A 16S database that has modified headers and is output in the reference_databases folder.
}
\description{
An ITS database that has modified headers and is output in the reference_databases folder.

An 16S database that has modified headers and is output in the reference_databases folder-Silva db is one that has species levels taxonomic assignments and based off the format of version 138.1.
}
\keyword{internal}
