% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_db_16S}
\alias{format_db_16S}
\title{An 16S database that has modified headers and is output in the
reference_databases folder}
\usage{
format_db_16S(
  data_tables,
  data_path,
  output_directory_path,
  temp_directory_path,
  db_16S
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, primer sequences}

\item{data_path}{Path to the data directory}

\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{temp_directory_path}{User-defined temporary directory to place reads
throughout the workflow metadata, and primer_info files}

\item{db_16S}{The name of the database}
}
\value{
An 16S database that has modified headers and is output in the
reference_databases folder
}
\description{
An 16S database that has modified headers and is output in the
reference_databases folder
}
\keyword{internal}
