% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_db_16s}
\alias{format_db_16s}
\title{An 16s database that has modified headers and is output in the reference_databases folder.}
\usage{
format_db_16s(analysis_setup, db_16s)
}
\arguments{
\item{db_16s}{The name of the database}

\item{directory_path}{The path to the directory containing the fastq,
metadata, and primer_info files}

\item{directory_path_temp}{User-defined temporary directory to place reads throughout the workflow
metadata, and primer_info files}
}
\value{
An 16s database that has modified headers and is output in the reference_databases folder.
}
\description{
An 16s database that has modified headers and is output in the reference_databases folder.
}
\keyword{internal}
