% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_db_other1}
\alias{format_db_other1}
\title{An other, user-specified database that is initially in the format specified by DADA2 with header simply taxonomic levels (kingdom down to species, separated by semi-colons, ;)}
\usage{
format_db_other1(analysis_setup, db_other1)
}
\arguments{
\item{db_other1}{The name of the database}

\item{directory_path}{The path to the directory containing the fastq,
metadata, and primer_info files}

\item{directory_path_temp}{User-defined temporary directory to place reads throughout the workflow
metadata, and primer_info files}
}
\value{
An other1 database that has modified headers and is output in the reference_databases folder.
}
\description{
An other, user-specified database that is initially in the format specified by DADA2 with header simply taxonomic levels (kingdom down to species, separated by semi-colons, ;)
}
\keyword{internal}
