% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_databases.R
\name{format_db_other2}
\alias{format_db_other2}
\title{An second user-specified database that is initially in the format specified
by dada2 with header containing taxonomic levels (kingdom down to species, separated by semi-colons)}
\usage{
format_db_other2(
  data_tables,
  data_path,
  output_directory_path,
  temp_directory_path,
  db_other2
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, and metabarcode information with associated primer sequences}

\item{data_path}{Path to the data directory}

\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{temp_directory_path}{User-defined temporary directory to output unfiltered, trimmed, and filtered read directories throughout the workflow}

\item{db_other2}{A second reference database other than SILVA, UNITE, or oomyceteDB (assumes format is like SILVA DB entries)}
}
\value{
The database with modified headers
}
\description{
An second user-specified database that is initially in the format specified
by dada2 with header containing taxonomic levels (kingdom down to species, separated by semi-colons)
}
\keyword{internal}
