% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{get_pids}
\alias{get_pids}
\title{Align ASV sequences to reference sequences from database to get percent ID. Get percent identities.}
\usage{
get_pids(tax_results, db)
}
\arguments{
\item{tax_results}{The dataframe containing taxonomic assignments}
}
\description{
Align ASV sequences to reference sequences from database to get percent ID. Get percent identities.
}
\keyword{internal}
