% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutPrimers_TrimReads.R
\name{get_post_primer_hits}
\alias{get_post_primer_hits}
\title{Get primer counts fo reach sample after primer removal and trimming steps}
\usage{
get_post_primer_hits(primer_data, cutadapt_data, directory_path)
}
\arguments{
\item{primer_data}{The primer data tibble created in orient_primers function}

\item{cutadapt_data}{directory_data folder with trimmed and filtered reads for each sample}
}
\value{
Table of read counts acros each sample
}
\description{
Get primer counts fo reach sample after primer removal and trimming steps
}
\keyword{internal}
