% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_primers_trim_reads.R
\name{get_post_trim_hits}
\alias{get_post_trim_hits}
\title{Get primer counts for reach sample after primer removal and trimming steps}
\usage{
get_post_trim_hits(primer_data, cutadapt_data, directory_path)

get_post_trim_hits(primer_data, cutadapt_data, directory_path)
}
\arguments{
\item{primer_data}{The primer data frame created in orient_primers function}

\item{cutadapt_data}{directory_data folder with trimmed and filtered reads for each sample}

\item{directory_path}{}
}
\value{
Table of read counts across each sample

Table of read counts across each sample
}
\description{
Get primer counts for reach sample after primer removal and trimming steps

Get primer counts for reach sample after primer removal and trimming steps
}
\keyword{internal}
