% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reads_count_primers.R
\name{get_pre_primer_hits}
\alias{get_pre_primer_hits}
\title{Get primer counts for reach sample before primer removal and trimming steps}
\usage{
get_pre_primer_hits(primer_data, fastq_data, output_directory_path)
}
\arguments{
\item{primer_data}{The primer data data frame created in orient_primers
function}

\item{fastq_data}{A data frame with the FASTQ file paths, the direction of
the sequences, and names of sequences}

\item{output_directory_path}{The path to the directory where resulting files
are output}
}
\value{
The number of reads in which the primer is found

The number of reads in which the primer is found
}
\description{
Get primer counts for reach sample before primer removal and trimming steps
}
\keyword{Get}
\keyword{and}
\keyword{before}
\keyword{counts}
\keyword{for}
\keyword{internal}
\keyword{primer}
\keyword{reach}
\keyword{removal}
\keyword{sample}
\keyword{steps}
\keyword{trimming}
