% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{get_ref_seq}
\alias{get_ref_seq}
\title{Align ASV sequences to reference sequences from database to get percent ID. STart by retrieving reference sequences.}
\usage{
get_ref_seq(tax_results, db)
}
\description{
Align ASV sequences to reference sequences from database to get percent ID. STart by retrieving reference sequences.
}
\keyword{internal}
