% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asv_abund_matrix.R
\name{infer_asvs}
\alias{infer_asvs}
\title{Core 'DADA2' function to learn errors and infer ASVs}
\usage{
infer_asvs(
  data_tables,
  my_direction,
  my_primer_pair_id,
  barcode_params,
  output_directory_path
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, and metabarcode information with associated primer sequences}

\item{my_direction}{Whether primer is in forward or reverse direction}

\item{my_primer_pair_id}{The specific metabarcode ID}

\item{output_directory_path}{The path to the directory where resulting files are output}
}
\value{
asv_data
}
\description{
Core 'DADA2' function to learn errors and infer ASVs
}
\keyword{internal}
