% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeASVAbundMatrix.R
\name{make_asv_abund_matrix}
\alias{make_asv_abund_matrix}
\title{Make an Amplified Sequence Variant (ASV) Abundance Matrix
This function generates an ASV abundance matrix based on data processed through
the preceding steps, including read preparation, cut and trim, and ASV inference.}
\usage{
make_asv_abund_matrix(analysis_setup, overwrite_existing = FALSE)
}
\arguments{
\item{analysis_setup}{A list containing directory paths and data tables, produced by the
\code{prepare_reads} function.}

\item{overwrite_existing}{Logical, indicating whether to overwrite existing results.}
}
\value{
The ASV abundance matrix (\code{asv_abund_matrix}).
}
\description{
Make an Amplified Sequence Variant (ASV) Abundance Matrix
This function generates an ASV abundance matrix based on data processed through
the preceding steps, including read preparation, cut and trim, and ASV inference.
}
\details{
The function processes data for each unique barcode separately, inferring
ASVs, merging reads, and creating an ASV abundance matrix.
}
\examples{
The primary wrapper function for DADA2 ASV inference steps
prepare_reads(maxN = 0, data_directory = "~/demulticoder/inst/extdata", output_directory = "~/testing_package", tempdir_id = "run1", overwrite_existing = TRUE)
cut_trim(analysis_setup,cutadapt_path="/opt/homebrew/bin/cutadapt", overwrite_existing = TRUE)
make_asv_abund_matrix(analysis_setup, overwrite_existing = TRUE)
}
