% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{make_cutadapt_tibble}
\alias{make_cutadapt_tibble}
\title{Prepare for primmer trimming with Cutaapt. Make new sub-directories
and specify paths for the trimmed and untrimmed reads}
\usage{
make_cutadapt_tibble(fastq_data, metadata, directory_path_temp)
}
\arguments{
\item{fastq_data}{A path to FASTQ files for analysis
metadata, and primer_info files}

\item{metadata}{A metadata containing the concatenated metadata and primer data}

\item{directory_path_temp}{User-defined temporary directory to place reads throughout the workflow.}
}
\value{
Returns a tibble that is used as input when running Cutadapt
}
\description{
Prepare for primmer trimming with Cutaapt. Make new sub-directories
and specify paths for the trimmed and untrimmed reads
}
\keyword{internal}
