% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CutPrimers_TrimReads.R,
%   R/PrepareReads_CountPrimers.R
\name{make_primer_hit_plot}
\alias{make_primer_hit_plot}
\title{Make a barplot of primers identified on reads}
\usage{
make_primer_hit_plot(primer_hits, fastq_data, directory_path, plot_name)

make_primer_hit_plot(primer_hits, fastq_data, directory_path, plot_name)
}
\arguments{
\item{primer_hits}{A number of reads in which the primer is found}

\item{fastq_data}{A tibble with the fastq file paths, the direction of
the sequences, and names of sequences}

\item{directory_path}{Folder path of read files and metadata file}

\item{plot_name}{A filename under which a PDF file of the plot will be saved as}
}
\value{
Make a plot showing any primers that may remain on reads

Returns a barplot with read counts
}
\description{
Make a barplot of primers identified on reads

Make a barplot of primers identified on reads
}
\keyword{internal}
