% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asv_abund_matrix.R
\name{make_seqhist}
\alias{make_seqhist}
\title{Plots a histogram of read length counts of all sequences within the ASV matrix}
\usage{
make_seqhist(asv_abund_matrix, output_directory_path)
}
\arguments{
\item{asv_abund_matrix}{The returned final ASV abundance matrix}
}
\value{
histogram with read length counts of all sequences within ASV matrix
}
\description{
Plots a histogram of read length counts of all sequences within the ASV matrix
}
\keyword{internal}
