% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeASVAbundMatrix.R
\name{makeSeqtab}
\alias{makeSeqtab}
\title{Make ASV sequence matrix}
\usage{
makeSeqtab(merged_reads, orderBy = "abundance")
}
\arguments{
\item{orderBy}{(Optional). \code{character(1)}. Default "abundance".
Specifies how the sequences (columns) of the returned table should be ordered (decreasing).
Valid values: "abundance", "nsamples", NULL.}
}
\value{
raw_seqtab
}
\description{
Make ASV sequence matrix
}
\keyword{internal}
