% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_asv_abund_matrix.R
\name{merge_reads_command}
\alias{merge_reads_command}
\title{Merge forward and reverse reads}
\usage{
merge_reads_command(
  output_directory_path,
  temp_directory_path,
  barcode_params,
  barcode
)
}
\arguments{
\item{output_directory_path}{The path to the directory where resulting files are output}

\item{merged_read_data_path}{Path to R data file containing merged read data}
}
\value{
merged_reads Intermediate merged read R data file
}
\description{
Merge forward and reverse reads
}
\keyword{internal}
