% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reads_count_primers.R
\name{orient_primers}
\alias{orient_primers}
\title{Take in user's forward and reverse sequences and creates the complement,
reverse, reverse complement of primers in one data frame}
\usage{
orient_primers(primers_params_path)
}
\arguments{
\item{primers_params_path}{A path to the CSV file that holds the primer
information.}
}
\value{
A data frame with oriented primer information.
}
\description{
Take in user's forward and reverse sequences and creates the complement,
reverse, reverse complement of primers in one data frame
}
\keyword{internal}
