% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_primers_trim_reads.R
\name{plot_post_trim_qc}
\alias{plot_post_trim_qc}
\title{Wrapper script for plotQualityProfile after trim steps and primer removal.}
\usage{
plot_post_trim_qc(
  cutadapt_data,
  output_directory_path,
  n = 5e+05,
  barcode_params
)
}
\arguments{
\item{cutadapt_data}{FASTQ read files trimmed of primers}

\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}
}
\value{
Quality profiles of reads after primer trimming
}
\description{
Wrapper script for plotQualityProfile after trim steps and primer removal.
}
\keyword{internal}
