% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_primers_trim_reads.R
\name{plot_qc}
\alias{plot_qc}
\title{Wrapper function for plotQualityProfile function}
\usage{
plot_qc(cutadapt_data, output_directory_path, n = 5e+05, barcode_params)
}
\arguments{
\item{cutadapt_data}{FASTQ read files trimmed of primers}

\item{output_directory_path}{The path to the directory where resulting files
are output}

\item{n}{(Optional). Default 500,000.
The number of records to sample from the fastq file.}

\item{seed}{Set seed for reproducibility}
}
\value{
Dada2 wrapper function for making quality profiles for each sample
}
\description{
Wrapper function for plotQualityProfile function
}
\keyword{internal}
