% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTaxonomy.R
\name{prep_abund_matrix}
\alias{prep_abund_matrix}
\title{Prepare final ASV abundance matrix}
\usage{
prep_abund_matrix(cutadapt_data, asv_abund_matrix, locus)
}
\arguments{
\item{directory_data}{folder with trimmed and filtered reads for each sample}
}
\description{
Prepare final ASV abundance matrix
}
\keyword{internal}
