% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{prepare_metadata_table}
\alias{prepare_metadata_table}
\title{Read in the metadata from user and combine it with the primer data.
Included in a larger function prepare_reads.}
\usage{
prepare_metadata_table(metadata_path, primer_data)
}
\arguments{
\item{metadata_path}{The path to the metadata file}

\item{primer_path}{a path to the csv file that holds the primer
information}
}
\value{
A dataframe containing the merged metadata and primer
data
}
\description{
Read in the metadata from user and combine it with the primer data.
Included in a larger function prepare_reads.
}
\keyword{internal}
