% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reads_count_primers.R
\name{prepare_metadata_table}
\alias{prepare_metadata_table}
\title{Read metadata file from user and combine and reformat it, given primer data.
Included in a larger function prepare_reads.}
\usage{
prepare_metadata_table(metadata_file_path, primer_data)
}
\arguments{
\item{primer_data}{A data frame of oriented primer information returned from the orient_primers function.}

\item{metadata_path}{The path to the metadata file.}
}
\value{
A dataframe containing the merged metadata and primer data.
}
\description{
Read metadata file from user and combine and reformat it, given primer data.
Included in a larger function prepare_reads.
}
