% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{prepare_reads}
\alias{prepare_reads}
\title{Prepare reads for primer trimming using Cutadapt}
\usage{
prepare_reads(
  maxN = 0,
  multithread = FALSE,
  overwrite_existing = FALSE,
  data_directory = "path/to/data",
  output_directory = "path/to/output",
  tempdir_id = "run1"
)
}
\arguments{
\item{maxN}{Integer, maximum number of Ns allowed in the sequence. Sequences with more than this many Ns will be discarded. Default is 0.}

\item{multithread}{Logical, indicating whether multithreading should be used. Default is FALSE.}

\item{overwrite_existing}{Logical, indicating whether to remove or overwrite existing files and directories from previous runs. If set to TRUE, specific output files will be deleted if they exist, ensuring a fresh start for the current run. Default is TRUE.}

\item{data_directory}{Character string specifying the directory where data files are located. Default is "path/to/data".}

\item{output_directory}{Character string specifying the directory for analysis outputs. Default is "path/to/output".}

\item{tempdir_id}{Character string specifying an ID for temporary directories. This ID will be combined with the current date. Default is "run1".}
}
\value{
A list with the following elements:
\itemize{
\item \code{data_tables}: A list containing data tables related to Cutadapt trimming, primer sequences, input FASTQ files, and metadata.
\item \code{dir_paths}: A list containing the paths set up by the \code{setup_directories} function, including data directory, output directory, temporary directory, primer path, and metadata path.
}
}
\description{
This function prepares sequencing reads for primer trimming using Cutadapt by setting up necessary directories, reading in metadata and primer data, and performing pre-trimming quality control steps.
}
\examples{
\dontrun{
outputs <- prepare_reads(maxN = 0, data_directory = "path/to/data", output_directory = "path/to/output")
}

}
