% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reads_count_primers.R
\name{prepare_reads}
\alias{prepare_reads}
\title{Prepare reads for primer trimming using Cutadapt}
\usage{
prepare_reads(
  data_directory = "data",
  output_directory = "output",
  tempdir_path = NULL,
  tempdir_id = "demulticoder_run",
  multithread = FALSE,
  overwrite_existing = FALSE
)
}
\arguments{
\item{data_directory}{User-specified directory path where user has placed raw FASTQ (forward and reverse reads), metadata.csv, and primerinfo_params.csv files. Default is "data".}

\item{output_directory}{User-specified directory for outputs. Default is "output".}

\item{tempdir_path}{Path to a temporary directory. If NULL, a temporary directory path will be identified using tempdir() command.}

\item{tempdir_id}{ID for temporary directories. Default is "demulticoder_run". User can provide any helpful ID, whether it be a date or specific name for the run.}

\item{multithread}{(Optional). Default is FALSE.
 If TRUE, input files are filtered in parallel via \code{\link[parallel]{mclapply}}.
 If an integer is provided, it is passed to the \code{mc.cores} argument of \code{\link[parallel]{mclapply}}.
 Note that the parallelization here is by forking, and each process is loading another fastq file into
 memory. This option is ignored in Windows, as Windows does not support forking, with \code{mc.cores} set to 1.
If memory is an issue, execute in a clean environment and reduce the chunk size \code{n} and/or
 the number of threads.}

\item{overwrite_existing}{Logical, indicating whether to remove or overwrite existing files and directories from previous runs.}
}
\value{
A list containing data tables, including metadata, primer sequences, paths for trimming reads, and user-defined parameters.
}
\description{
This function organizes user-defined directories, metadata and parameter inputs in preparation for Cutadapt to remove primers from short read metabarcoding sequence data. Afterwards, subsequent functions can be used to simplify core DADA2 ASV and taxonomic assignment steps.
}
\examples{
Pre-filter raw reads and parse metadata and primer_information to prepare for primer trimming and filter
prepare_reads(data_directory = "~/demulticoder/inst/extdata", output_directory = "~/testing_package", tempdir_id = "demulticoder_run", overwrite_existing = TRUE)
}
