% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_reads_count_primers.R
\name{primer_check}
\alias{primer_check}
\title{Matching Order Primer Check}
\usage{
primer_check(fastq_data)
}
\arguments{
\item{fastq_data}{A data frame containing the read file paths and the direction of the reads by sample}
}
\value{
None
}
\description{
Matching Order Primer Check
}
\keyword{internal}
