% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{process_single_barcode}
\alias{process_single_barcode}
\title{Process the information from an ASV abundance matrix to run DADA2 for single
barcode}
\usage{
process_single_barcode(
  data_tables,
  temp_directory_path,
  output_directory_path,
  asv_abund_matrix,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = FALSE,
  locus = barcode
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, primer sequences}

\item{temp_directory_path}{The temporary directory path}

\item{asv_abund_matrix}{An abundance matrix containing amplified sequence
variants}

\item{tryRC}{Try reverse complement (default FALSE)}

\item{verbose}{Print additional information (default FALSE)}

\item{multithread}{Use multiple threads (default TRUE)}

\item{locus}{The locus for taxonomy assignment (e.g., rps10, other1, other2)}
}
\description{
Process the information from an ASV abundance matrix to run DADA2 for single
barcode
}
\keyword{internal}
