% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_taxonomy.R
\name{process_single_barcode}
\alias{process_single_barcode}
\title{Process the information from an ASV abundance matrix to run DADA2 for single
barcode}
\usage{
process_single_barcode(
  data_tables,
  temp_directory_path,
  output_directory_path,
  asv_abund_matrix,
  tryRC = FALSE,
  verbose = FALSE,
  multithread = FALSE,
  locus = barcode
)
}
\arguments{
\item{data_tables}{The data tables containing the paths to read files, metadata, primer sequences}

\item{asv_abund_matrix}{An abundance matrix containing amplified sequence
variants}
}
\description{
Process the information from an ASV abundance matrix to run DADA2 for single
barcode
}
\keyword{internal}
