% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{read_fastq}
\alias{read_fastq}
\title{Takes in the fastq files from the user and creates a tibble with
the paths to files that will be created and used in the future.
Included in a larger 'read_prefilt_fastq' function}
\usage{
read_fastq(directory_path, directory_path_temp)
}
\arguments{
\item{directory_path}{The path to the directory containing the fastq,
metadata, and primer_info files}

\item{directory_path_temp}{User-defined temporary directory to place reads throughout the workflow.}
}
\value{
A tibble with the fastq file paths, the direction of the
}
\description{
Takes in the fastq files from the user and creates a tibble with
the paths to files that will be created and used in the future.
Included in a larger 'read_prefilt_fastq' function
}
\keyword{and}
\keyword{internal}
\keyword{names}
\keyword{of}
\keyword{sequences}
\keyword{sequences,}
