% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DADA2_wrapper_functions.R
\name{separate_abund_matrix}
\alias{separate_abund_matrix}
\title{Still needs improvement. Could solve problem by merging fungal and oomycete databases for assign taxonomy steps. In order to assign taxonomy for separate loci, need to separate abundance tables first}
\usage{
separate_abund_matrix(
  abund_asv_barcode2,
  abund_asv_barcode1,
  directory_path,
  asv_abund_matrix
)
}
\arguments{
\item{directory_path}{A path to the intermediate folder and directory}

\item{asv_abund_matrix}{}
}
\description{
Still needs improvement. Could solve problem by merging fungal and oomycete databases for assign taxonomy steps. In order to assign taxonomy for separate loci, need to separate abundance tables first
}
