% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepareReads_CountPrimers.R
\name{setup_directories}
\alias{setup_directories}
\title{Set Up Directory Paths for Analysis}
\usage{
setup_directories(
  data_directory = "data",
  output_directory = "outputs",
  tempdir_id = "run1"
)
}
\arguments{
\item{data_directory}{Character string specifying the directory where data files are located.
Default is "inst/extdata".}

\item{output_directory}{Character string specifying the directory for analysis outputs.
Default is "outputs".}

\item{tempdir_id}{Character string specifying an ID for temporary directories.
This ID will be combined with the current date. Default is "run1".}
}
\value{
A list with the following elements:
\itemize{
\item \code{data_directory}: The data directory path.
\item \code{output_directory}: The output directory path.
\item \code{temp_directory}: The path to the temporary directory.
\item \code{primer_path}: The path to the primer information file.
\item \code{metadata_path}: The path to the metadata file.
}
\description{
This function sets up the paths for the analysis.
It checks whetherthe specified output directories exist or creates them if they don't.
The function also provides paths to primer and metadata files within the data directory.
}
\examples{
\dontrun{
dirs <- setup_directories(data_directory = "path/to/my/data", 
                          output_directory = "path/to/output")
print(dirs$primer_path)
}


}
