% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConvertASVMatrix.R
\name{taxmap_to_phyloseq}
\alias{taxmap_to_phyloseq}
\title{Convert taxmap object to Phyloseq object (metacoder wrapper function)}
\usage{
taxmap_to_phyloseq(obj_dada)
}
\arguments{
\item{obj_data}{ASV matrix converted to taxmap object using asvmatrix_to_taxmap function}
}
\value{
Taxmap object converted to a phyloseq object
}
\description{
Convert taxmap object to Phyloseq object (metacoder wrapper function)
}
\examples{
directory_path<-"~/Desktop/rps10test"
primer_path <-file.path(directory_path, "primer_info.csv")
metadata_path <-file.path(directory_path,"metadata.csv")
cutadapt_path<-"/opt/homebrew/bin/cutadapt"
data_tables <-
prepare_reads(
directory_path,
primer_path,
metadata_path,
maxN = 0,
)
cut_trim(
directory_path,
cutadapt_path,
verbose = TRUE,
maxEE = 2,
truncQ = 5,
minLen = 200,
maxLen = 297,
minCutadaptlength = 50
)
asv_abund_matrix <-
make_asv_abund_matrix(
directory_path,
minOverlap = 15,
maxMismatch = 2,
verbose = TRUE,
)
summary <- assignTax(
directory_path,
data_tables,
asv_abund_matrix,
multithread = TRUE,
barcode = "rps10",
database_rps10 = "oomycetedb.fasta",
database_its = "sh_general_release_dynamic_22.08.2016.fasta"
)
obj_dada<-asvmatrix_to_taxmap(
asv_abund_matrix,
min_read_depth=10,
minimum_bootstrap=75
)
phylo_obj<-taxmap_to_phyloseq(obj_dada)
}
